@@ -1,9 +1,13 @@
 import { createAction, props } from "@ngrx/store";
 import { IUser } from "../models/IUser";
-import {LOGIN_ACTION} from './types'
+import {LOGIN_ACTION, LOGOUT_ACTION} from './types'
 
 
 export const login = createAction(
     LOGIN_ACTION,
-    props<IUser>()
-)
+    props<{user: IUser}>()
+);
+
+export const logout = createAction(
+    LOGOUT_ACTION
+);
