@@ -1,6 +1,8 @@
 export interface IUser {
-    username: string,
-    email: string,
-    password: string,
-    gender: string
+    id?: number,
+    username?: string,
+    email?: string,
+    password?: string,
+    gender?: string,
+    role?: string;
 }
