@@ -1,4 +1,11 @@
+import { state } from '@angular/animations';
 import { Component, OnInit } from '@angular/core';
+import { select, Store } from '@ngrx/store';
+import { IUser } from 'app/models/IUser';
+import { UserState } from 'app/reducers/auth.reducer';
+import { loggedUser } from 'app/selectors/login.selector';
+import { map, Observable } from 'rxjs';
+import * as AuthActions from '../../actions/auth.actions';
 
 @Component({
   selector: 'app-navbar',
@@ -11,10 +18,20 @@ export class NavbarComponent implements OnInit {
   readonly loginButton = 'Login';
   readonly registerButton = 'Register';
   readonly logoutButton = 'Logout';
-
-  constructor() { }
+  currentUser$: Observable<IUser> | undefined;
+  
+  constructor(private store: Store<UserState>) {
+    
+   }
 
   ngOnInit(): void {
+    this.currentUser$ = this.store.pipe(
+      select(loggedUser)
+    );
+  }
+
+  logout() {
+    this.store.dispatch(AuthActions.logout())
   }
 
 }
